+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 2.1.3 (Next)
 **AWX**
 - Add `instanceId` parameter to `hostCreate` method.
+- Update `smart` parameter to `kind` for inventory methods.
 
 **GoSS**
 - Add `logLevel` parameter to `server` and `validate` methods.